[REFACTOR] PostEntity내부의 게시판 eunm을 외부로 분리
