+++ src/main/java/com/kakaobase/snsapp/domain/posts/entity/Post.java
@@ -1,6 +1,7 @@
 package com.kakaobase.snsapp.domain.posts.entity;
 
 import com.kakaobase.snsapp.domain.members.entity.Member;
+import com.kakaobase.snsapp.domain.posts.util.BoardType;
 import com.kakaobase.snsapp.global.common.entity.BaseSoftDeletableEntity;
 import jakarta.persistence.*;
 import lombok.*;
@@ -66,15 +67,6 @@ public class Post extends BaseSoftDeletableEntity {
         this.boardType = boardType;
         this.content = content;
         this.youtubeUrl = youtubeUrl;
-    }
-
-    public enum BoardType {
-        ALL,        // 전체 게시판
-        PANGYO_1,   // 판교 1기 게시판
-        PANGYO_2,   // 판교 2기 게시판
-        JEJU_1,     // 제주 1기 게시판
-        JEJU_2,     // 제주 2기 게시판
-        JEJU_3      // 제주 3기 게시판
     }
 
     /**
