+++ src/main/java/com/kakaobase/snsapp/domain/posts/util/BoardType.java
@@ -1,0 +1,10 @@
+package com.kakaobase.snsapp.domain.posts.util;
+
+public enum BoardType {
+        ALL,        // 전체 게시판
+        PANGYO_1,   // 판교 1기 게시판
+        PANGYO_2,   // 판교 2기 게시판
+        JEJU_1,     // 제주 1기 게시판
+        JEJU_2,     // 제주 2기 게시판
+        JEJU_3      // 제주 3기 게시판
+}
